refactor(run-once): extract structured result summary
